Increased max width size to 200.
Thinking of uncapping the width.
Also fixed the progress bar. Was another case of int division resulting in 0. So it was 0 for 99% of the bar and 100% when count and total were the same.

+++ DMCConverter/ConvertImage.cs
@@ -98,7 +98,7 @@ namespace DMCConverter
 
                     //increase the value of the progress bar
                     count += 1;
-                    progressBar.Value = Convert.ToInt32(Math.Round((count / total)*100f));
+                    progressBar.Value = Convert.ToInt32(Math.Round(((float)count / (float)total)*100f));
